models: Move model into their own directory
This allows there to be a file that is a list of models that is
not mixed into the runner code.
