+++ model/models/mllama/process_image.go
