+++ model/models/pixtral/imageproc.go
