+++ model/models/pixtral/imageproc_test.go
