+++ model/models/qwen2vl/imageproc.go
