+++ model/models/qwen2vl/imageproc_test.go
